test_evens.py
import unittest
from evens import even_number_of_evens


class TestEvens(unittest.TestCase):
    def test_throws_error_if_value_passed_in_is_not_list(self):
        self.assertRaises(TypeError, even_number_of_evens, 4)
    
    def test_values_in_list(self):
        self.assertEqual(even_number_of_evens([]), False)
        self.assertEqual(even_number_of_evens([2, 4]), True)
        self.assertEqual(even_number_of_evens([2]), False)
        self.assertEqual(even_number_of_evens([1, 3, 5]), False)
        
        """
        The assertRaises() method is used in unit testing to verify that a specific exception
        is raised when a certain piece of code is executed. It is commonly used in Python's
        unittest framework.
        """


if __name__ == "__main__":
    unittest.main()